pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "image-search"
version = "0.1.0"
description = "CLI tool for semantic image search using CLIP embeddings and sqlite-vec"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "numpy<2",
    "openai-clip>=1.0.1",
    "pillow>=11.2.1",
    "sentence-transformers>=4.1.0",
    "setuptools>=80.9.0",
    "sqlite-vec>=0.1.6",
    "torch==2.2.0",
    "torchvision>=0.17.0",
    "transformers>=4.53.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
image-search = "image_search_cli:main"

[tool.hatch.build.targets.wheel]
only-include = ["image_search_cli.py"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
    "--strict-markers",
    "--disable-warnings",
    "-v"
]
markers = [
    "unit: Fast unit tests with mocked dependencies",
    "integration: Slower integration tests with real models",
    "slow: Very slow tests that download models"
]

[dependency-groups]
test = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
]
